src/postprocessing.py
from datetime import datetime
from dateutil.relativedelta import relativedelta
import json

from .nlp.transliteration import transliterate_ukrainian_to_english

from elastic.business_search import search_businesses
from logging_config import logger


def process_time_llm_response(
    response, current_time=datetime.now().isoformat()
):
    """Process time response from LLM and convert it to datetime format."""
    try:
        data = json.loads(response)
        logger.debug(f"Data structure from LLM: {data}")
        logger.debug(f"Current time: {current_time}")

        # Convert current_time to datetime object
        if isinstance(current_time, str):
            current_time_dt = datetime.fromisoformat(
                current_time.replace("Z", "+00:00")
            )
        else:
            current_time_dt = current_time

        if data["time"]:
            return {"datetime": data["time"]}
        else:
            if data["action"] == "+":
                time_change = relativedelta(
                    years=data["years"] or 0,
                    months=data["months"] or 0,
                    days=data["days"] or 0,
                    hours=data["hours"] or 0,
                    minutes=data["minutes"] or 0,
                )
                new_time = current_time_dt + time_change
            elif data["action"] == "-":
                time_change = relativedelta(
                    years=-(data["years"] or 0),
                    months=-(data["months"] or 0),
                    days=-(data["days"] or 0),
                    hours=-(data["hours"] or 0),
                    minutes=-(data["minutes"] or 0),
                )
                new_time = current_time_dt + time_change
            logger.debug(f"Processed time: {new_time.isoformat()}")
            return {"datetime": new_time.isoformat()}

    except Exception as e:
        logger.error(f"Error processing time: {e}", exc_info=True)
        # Fall back to the provided reference time to keep evaluation deterministic
        try:
            ref_time = (
                current_time
                if isinstance(current_time, str)
                else current_time.isoformat()
            )
        except Exception:
            ref_time = datetime.now().isoformat()
        return {"datetime": str(ref_time)}


def process_business_llm_response(response):
    """Process business response from LLM and search for matching businesses."""
    try:
        data = json.loads(response)
        logger.debug(f"Extracted business by LLM: {data.get('business')}")
        logger.debug(f"LLM response: {data}")

        if not data.get("business"):
            logger.debug("No business entity extracted by LLM.")
            return {"business_id": None}

        # Handle Ukrainian business input
        if data.get("language") == "uk":
            data["orig_uk_to_en_transliteration"] = (
                transliterate_ukrainian_to_english(data["business"])
            )
            data["lemma_uk_to_en_transliteration"] = (
                transliterate_ukrainian_to_english(data.get("uk_lemma", ""))
            )

        # Prepare values for search: remove 'language' and empty values
        filtered_data = {
            k: v for k, v in data.items() if k != "language" and v
        }

        try:
            search_values = set(filtered_data.values())
            logger.debug(f"Searching businesses by values: {search_values}")
            matched_businesses = search_businesses(search_values)
        except Exception as search_exc:
            logger.error(f"Error searching businesses: {search_exc}")
            matched_businesses = None

        logger.debug(f"Matched businesses found: {matched_businesses}")

        if matched_businesses:
            return {"businesses": matched_businesses}

        logger.info("No matched businesses found.")
        return {"business_id": None}
    except Exception as e:
        logger.error(
            f"Error processing business LLM response: {e}", exc_info=True
        )
        return {"business_id": None}
