requirements.txt
dotenv==0.9.9
httpx==0.28.1
functions-framework==3.8.2
python-dateutil==2.9.0
elasticsearch==8.15.0
transliterate==1.10.2
black==24.8.0
isort==5.13.2
